# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  controllers.PromotionRequestResponse:
    properties:
      admin_notes:
        type: string
      created_at:
        type: string
      id:
        type: string
      plan_id:
        type: string
      reviewed_at:
        type: string
      status:
        type: string
      user_id:
        type: string
    type: object
...
  controllers.SubmitPromotionRequest:
    properties:
      plan_id:
        description: 'Optional: if provided, request is for plan promotion; if null,
          request is for user role upgrade'
        type: string
    type: object
  controllers.SubmitRatingRequest:
    properties:
      stars:
        maximum: 5
        minimum: 1
        type: integer
    required:
    - stars
    type: object
...
  middleware.ErrorInfo:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
...
  middleware.SwaggerErrorResponse:
    properties:
      error:
        $ref: '#/definitions/middleware.ErrorInfo'
      success:
        example: false
        type: boolean
    type: object
host: localhost:8000
info:
  contact:
    email: support@travellink.local
    name: API Support
    url: http://www.swagger.io/support
  description: Travel Linked List API - Browse, create, and share travel plans
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: TravelLink API
  version: "1.0"
paths:
...
  /plans/{id}/my-rating:
    get:
      description: Retrieve your current rating for a travel plan. Returns null if
        not rated.
      parameters:
      - description: Plan ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Your rating or null
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
        "404":
          description: Plan not found
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
      security:
      - Bearer: []
      summary: Get your rating for a travel plan
      tags:
      - ratings
...
  /plans/{id}/ratings:
    get:
      description: Retrieve aggregate rating statistics for a travel plan (public
        endpoint)
      parameters:
      - description: Plan ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Rating statistics (average, count, sum)
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Plan not found
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
      summary: Get rating statistics for a travel plan
      tags:
      - ratings
    post:
      consumes:
      - application/json
      description: Rate a travel plan with 1-5 stars (upsert operation). Authenticated
        users can rate.
      parameters:
      - description: Plan ID
        in: path
        name: id
        required: true
        type: string
      - description: 'Rating submission request (stars: 1-5)'
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.SubmitRatingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Rating submitted with ID
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
        "404":
          description: Plan not found
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
      security:
      - Bearer: []
      summary: Submit a rating for a travel plan
      tags:
      - ratings
    put:
      consumes:
      - application/json
      description: Update your rating for a travel plan (1-5 stars). Authenticated
        users can update their rating.
      parameters:
      - description: Plan ID
        in: path
        name: id
        required: true
        type: string
      - description: 'Rating update request (stars: 1-5)'
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.SubmitRatingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Rating updated with ID
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
        "404":
          description: Plan not found
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
      security:
      - Bearer: []
      summary: Update a travel plan rating
      tags:
      - ratings
...
  /promotions/my-requests:
    get:
      description: Get all promotion requests submitted by the current user.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 50
        description: Results per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of promotion requests
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
      security:
      - Bearer: []
      summary: List my promotion requests
      tags:
      - promotions
  /promotions/request:
    post:
      consumes:
      - application/json
      description: Submit a request to upgrade to traveller role or for plan promotion.
        Simple users can submit requests.
      parameters:
      - description: Promotion request submission
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.SubmitPromotionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Request submitted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
        "404":
          description: Plan not found or user not found
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
      security:
      - Bearer: []
      summary: Submit a promotion request
      tags:
      - promotions
  /promotions/request/{id}:
    get:
      description: Get the status of a specific promotion request.
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Promotion request details
          schema:
            $ref: '#/definitions/controllers.PromotionRequestResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
        "404":
          description: Request not found
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
      security:
      - Bearer: []
      summary: Get promotion request status
      tags:
      - promotions
...
securityDefinitions:
  Bearer:
    description: 'Enter JWT token in the format: Bearer eyJhbGci...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"